Fix wExp in AdaptiveCurveIrmLib.ts
This implementation is confusing. When q is tested to be *equal* to zero, a left shift by 0 does nothing.

I suppose this is a port to javascript / typescript of this solidity code in https://github.com/morpho-org/morpho-blue-irm/blob/0e99e647c9bd6d3207f450144b6053cf807fa8c4/src/adaptive-curve-irm/libraries/ExpLib.sol#L45
```
// Return e^x = 2^q * e^r.
if (q >= 0) return expR << uint256(q);
else return expR >> uint256(-q);
```

Note the test is q >= 0, not q == 0.

In the current implementation, the right shift will be used, even if q is positive, i.e. -q is negative. This (probably) works because BigInt in javascript supports negative shifts.

However this is confusing when reading the code, and make it unobvious that this behaves like the original solidity code. Changing the test to the same form as the original solidity code fixes that.

Signed-off-by: harpocryptes <95710073+harpocryptes@users.noreply.github.com>